Reverted rename.
Reverted the rename of TokenStream and made
its constructor flexible in terms of the Stream
it is being made out of.
