--- pest/src/streams/mod.rs
@@ -1,1 +1,0 @@
-pub mod parser_stream;
