+++ pest/src/tokens/mod.rs
@@ -1,5 +1,7 @@
 mod token;
 mod token_data;
+mod token_stream;
 
 pub use self::token::Token;
 pub use self::token_data::TokenData;
+pub use self::token_stream::TokenStream;
